add: возможность изменить интервал постинга

+++ models.js
@@ -26,7 +26,7 @@ const DownloadQueueShema = sequalize.define('downloadQueue', {
 const InfoShema = sequalize.define('info', {
     ID: {type: DataTypes.INTEGER, primaryKey: true, unique: true, autoIncrement: true},
     countOfPost: {type: DataTypes.INTEGER},
-    postingInterval: {type: DataTypes.INTEGER},
+    postingInterval: {type: DataTypes.FLOAT},
     estimatedPostingTime: {type: DataTypes.STRING},
     onPosting: {type: DataTypes.BOOLEAN},
     userName: {type: DataTypes.STRING},
